# api/openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"description": ":phyor: :wade: :infoiasi: This is a restful api documentation implementation for Phyor Wade Project.",
		"version": "1.0.0",
		"title": "Phyor",
		"contact": {
			"email": "irinel.bucur@gmail.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "https://ibucur.zego.ro/api",
			"description": "Production server"
		}
	],
...
	"paths": {
		"/authors": {
			"post": {
				"tags": ["authors"],
				"summary": "Add a new author to the store",
				"description": "",
				"operationId": "addAuthor",
				"responses": {
					"201": {
						"$ref": "#/components/responses/Author"
					},
					"401": {
						"$ref": "#/components/responses/401"
					},
					"406": {
						"$ref": "#/components/responses/406"
					},
					"409": {
						"$ref": "#/components/responses/409"
					},
					"501": {
						"$ref": "#/components/responses/501"
					}
				},
				"security": [
					{
						"phyor_auth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Author"
							}
						}
					},
					"description": "Specifies the details of a new resource to be stored",
					"required": true
				},
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "The login provided authorization token",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "Accept",
						"in": "header",
						"description": "Specifies the output format. Only application/xml and application/json is supported. It another format is requested then the application/json format will be used.",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"tags": ["authors"],
				"summary": "Get all authors",
				"description": "",
				"operationId": "getAllAuthor",
				"responses": {
					"200": {
						"$ref": "#/components/responses/AuthorList"
					},
					"501": {
						"$ref": "#/components/responses/501"
					}
				},
				"parameters": [
					{
						"name": "Accept",
						"in": "header",
						"description": "Specifies the output format. Only application/xml and application/json is supported. It another format is requested then the application/json format will be used.",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "pageNumber",
						"in": "query",
						"description": "Specifies the page number to start showing the responses. If not specified, the default value is 0.",
						"required": false,
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "resultsPerPage",
						"in": "query",
						"description": "Specifies the number of resources which will be provided. If not specified, the default value is 20.",
						"required": false,
						"schema": {
							"type": "number"
						}
					}
				]
			}
		},
		"/authors/{authorId}": {
			"get": {
				"tags": ["authors"],
				"summary": "Get an author",
				"description": "",
				"operationId": "getAuthor",
				"responses": {
					"200": {
						"$ref": "#/components/responses/Author"
					},
					"404": {
						"$ref": "#/components/responses/404"
					},
					"501": {
						"$ref": "#/components/responses/501"
					}
				},
				"parameters": [
					{
						"name": "Accept",
						"in": "header",
						"description": "Specifies the output format. Only application/xml and application/json is supported. It another format is requested then the application/json format will be used.",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "authorId",
						"in": "path",
						"description": "Specifies the unique it of an author to get the details about.",
						"required": true,
						"schema": {
							"type": "number"
						}
					}
				]
			},
			"put": {
				"tags": ["authors"],
				"summary": "Update an author to the store",
				"description": "",
				"operationId": "saveAuthor",
				"responses": {
					"202": {
						"$ref": "#/components/responses/Author"
					},
					"401": {
						"$ref": "#/components/responses/401"
					},
					"406": {
						"$ref": "#/components/responses/406"
					},
					"409": {
						"$ref": "#/components/responses/409"
					},
					"501": {
						"$ref": "#/components/responses/501"
					}
				},
				"security": [
					{
						"phyor_auth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Author"
							}
						}
					},
					"description": "Specifies the details of a new resource to be stored",
					"required": true
				},
				"parameters": [
					{
						"name": "Authorization",
						"in": "header",
						"description": "The login provided authorization token",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "Accept",
						"in": "header",
						"description": "Specifies the output format. Only application/xml and application/json is supported. It another format is requested then the application/json format will be used.",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "authorId",
						"in": "path",
						"description": "Specifies the unique it of an author to get the details about.",
						"required": true,
						"schema": {
							"type": "number"
						}
					}
				]
			}
		},
...
	},
...
	"components": {
		"schemas": {
			"Author": {
				"description": "Author resource",
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int64",
						"description": "Unique identification number for the author"
					},
					"name": {
						"type": "string",
						"format": "varchar",
						"description": "Full name of the author"
					},
					"wikidataUri": {
						"type": "string",
						"format": "varchar",
						"description": "Specifies the external resource uri which better describes the author"
					},
					"resourceUri": {
						"type": "string",
						"format": "varchar",
						"description": "Specifies the local resource unique uri"
					}
				},
				"xml": {
					"name": "author"
				}
			},
			"AuthorList": {
				"description": "List of authors",
				"type": "array",
				"items": {
					"$ref": "#/components/schemas/Author"
				},
				"xml": {
					"name": "authors"
				}
			},
...
			"ApiErrorResponse": {
				"description": "API Error returned",
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"default": false,
						"description": "Specifies the success status of the operation"
					},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "Specifies an unique code to better identify the error"
					},
					"message": {
						"type": "string",
						"format": "varchar",
						"description": "Specifies a message which better describes the error"
					},
					"debug": {
						"type": "object",
						"properties": {
							"request": {
								"type": "object",
								"properties": {
									"request_url": {
										"type": "string",
										"description": "Specifies the called uri which generated the error"
									},
									"request_queryParams": {
										"type": "string",
										"description": "Specifies the query params used when called the uri which generated the error"
									},
									"request_uriParams": {
										"type": "string",
										"description": "Specifies the uri parameters like the book id or author id used when called the uri which generated the error"
									},
									"request_postParams": {
										"type": "object",
										"description": "Contains an object with all the values sent. It is different based on the post used."
									}
								}
							}
						}
					}
				}
			}
		},
...
		"responses": {
			"Author": {
				"description": "Author response",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Author"
						}
					},
					"application/xml": {
						"schema": {
							"$ref": "#/components/schemas/Author"
						}
					}
				}
			},
			"AuthorList": {
				"description": "Author list response",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/AuthorList"
						}
					},
					"application/xml": {
						"schema": {
							"$ref": "#/components/schemas/AuthorList"
						}
					}
				}
			},
...
			"401": {
				"description": "Unauthorized access",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					},
					"application/xml": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					}
				}
			},
			"404": {
				"description": "The requested resource does not exists",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					},
					"application/xml": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					}
				}
			},
			"406": {
				"description": "Submitted data validation failed",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					},
					"application/xml": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					}
				}
			},
			"409": {
				"description": "The submitted resource is not valid",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					},
					"application/xml": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					}
				}
			},
			"501": {
				"description": "Unexpected error",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					},
					"application/xml": {
						"schema": {
							"$ref": "#/components/schemas/ApiErrorResponse"
						}
					}
				}
			}
		},
		"securitySchemes": {
			"phyor_auth": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header"
			}
		}
	},
	"security": [

	]
}